pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bloxhub-discord-bot"
version = "1.0.0"
authors = [
    {name = "BloxHub Team", email = "example@example.com"}
]
description = "Discord bot for Roblox verification and gamepass management"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "discord.py>=2.0.0",
    "Flask>=2.0.0",
    "requests>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=0.19.0",
    "SQLAlchemy>=1.4.0",
    "APScheduler>=3.8.0",
    "gunicorn>=20.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "mypy>=0.950",
    "black>=22.3.0",
    "isort>=5.10.0",
]